[snapshot] Fix an unsigned-signed overflow in PutRaw
.. and related spots.

Fixed: 483851880
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7580174
Commit-Queue: Jakob Linke <jgruber@chromium.org>
Cr-Commit-Position: refs/heads/main@{#105279}

+++ src/snapshot/serializer.cc
@@ -463,7 +463,8 @@ void Serializer::ObjectSerializer::SerializePrologue(SnapshotSpace space,
     const char* code_name =
         serializer_->code_address_map_->Lookup(object_->address());
     LOG(serializer_->isolate_,
-        CodeNameEvent(object_->address(), sink_->Position(), code_name));
+        CodeNameEvent(object_->address(), static_cast<int>(sink_->Position()),
+                      code_name));
   }
 
   if (map.SafeEquals(*object_)) {
